Refactor logging and test environment into their own files.
