+++ wptrunner/wptrunner.py
@@ -5,30 +5,20 @@
 from __future__ import unicode_literals
 
 import json
-import logging
 import os
-import shutil
-import socket
 import sys
-import threading
-import time
-import urlparse
-from Queue import Empty
-from StringIO import StringIO
 
-from multiprocessing import Queue
-
-from mozlog.structured import (commandline, stdadapter, get_default_logger,
-                               structuredlog, handlers, formatters)
-
+import environment as env
 import products
 import testloader
 import wptcommandline
+import wptlogging
 import wpttest
 from testrunner import ManagerGroup
 
 here = os.path.split(__file__)[0]
 
+logger = None
 
 """Runner for web-platform-tests
 
@@ -46,266 +36,9 @@ format. This manifest is used directly to determine which tests exist. Local
 metadata files are used to store the expected test results.
 """
 
-logger = None
-
-
-def setup_logging(args, defaults):
+def setup_logging(*args, **kwargs):
     global logger
-    logger = commandline.setup_logging("web-platform-tests", args, defaults)
-    setup_stdlib_logger()
-
-    for name in args.keys():
-        if name.startswith("log_"):
-            args.pop(name)
-
-    return logger
-
-
-def setup_stdlib_logger():
-    logging.root.handlers = []
-    logging.root = stdadapter.std_logging_adapter(logging.root)
-
-
-def do_delayed_imports(serve_root):
-    global serve, sslutils
-
-    sys.path.insert(0, serve_root)
-
-    failed = []
-
-    try:
-        from tools.serve import serve
-    except ImportError:
-        failed.append("serve")
-
-    try:
-        import sslutils
-    except ImportError:
-        raise
-        failed.append("sslutils")
-
-    if failed:
-        logger.critical(
-            "Failed to import %s. Ensure that tests path %s contains web-platform-tests" %
-            (", ".join(failed), serve_root))
-        sys.exit(1)
-
-
-class TestEnvironmentError(Exception):
-    pass
-
-
-class LogLevelRewriter(object):
-    """Filter that replaces log messages at specified levels with messages
-    at a different level.
-
-    This can be used to e.g. downgrade log messages from ERROR to WARNING
-    in some component where ERRORs are not critical.
-
-    :param inner: Handler to use for messages that pass this filter
-    :param from_levels: List of levels which should be affected
-    :param to_level: Log level to set for the affected messages
-    """
-    def __init__(self, inner, from_levels, to_level):
-        self.inner = inner
-        self.from_levels = [item.upper() for item in from_levels]
-        self.to_level = to_level.upper()
-
-    def __call__(self, data):
-        if data["action"] == "log" and data["level"].upper() in self.from_levels:
-            data = data.copy()
-            data["level"] = self.to_level
-        return self.inner(data)
-
-
-class TestEnvironment(object):
-    def __init__(self, test_paths, ssl_env, options):
-        """Context manager that owns the test environment i.e. the http and
-        websockets servers"""
-        self.test_paths = test_paths
-        self.ssl_env = ssl_env
-        self.server = None
-        self.config = None
-        self.external_config = None
-        self.test_server_port = options.pop("test_server_port", True)
-        self.options = options if options is not None else {}
-        self.required_files = options.pop("required_files", [])
-        self.files_to_restore = []
-
-    def __enter__(self):
-        self.ssl_env.__enter__()
-        self.copy_required_files()
-        self.setup_server_logging()
-        self.setup_routes()
-        self.config = self.load_config()
-        serve.set_computed_defaults(self.config)
-        self.external_config, self.servers = serve.start(self.config, self.ssl_env)
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.ssl_env.__exit__(exc_type, exc_val, exc_tb)
-
-        self.restore_files()
-        for scheme, servers in self.servers.iteritems():
-            for port, server in servers:
-                server.kill()
-
-    def load_config(self):
-        default_config_path = os.path.join(serve_path(self.test_paths), "config.default.json")
-        local_config_path = os.path.join(here, "config.json")
-
-        with open(default_config_path) as f:
-            default_config = json.load(f)
-
-        with open(local_config_path) as f:
-            data = f.read()
-            local_config = json.loads(data % self.options)
-
-        #TODO: allow non-default configuration for ssl
-
-        local_config["external_host"] = self.options.get("external_host", None)
-        local_config["ssl"]["encrypt_after_connect"] = self.options.get("encrypt_after_connect", False)
-
-        config = serve.merge_json(default_config, local_config)
-        config["doc_root"] = serve_path(self.test_paths)
-
-        if not self.ssl_env.ssl_enabled:
-            config["ports"]["https"] = [None]
-
-        host = self.options.get("certificate_domain", config["host"])
-        hosts = [host]
-        hosts.extend("%s.%s" % (item[0], host) for item in serve.get_subdomains(host).values())
-        key_file, certificate = self.ssl_env.host_cert_path(hosts)
-
-        config["key_file"] = key_file
-        config["certificate"] = certificate
-
-        return config
-
-    def setup_server_logging(self):
-        server_logger = get_default_logger(component="wptserve")
-        assert server_logger is not None
-        log_filter = handlers.LogLevelFilter(lambda x:x, "info")
-        # Downgrade errors to warnings for the server
-        log_filter = LogLevelRewriter(log_filter, ["error"], "warning")
-        server_logger.component_filter = log_filter
-
-        try:
-            #Set as the default logger for wptserve
-            serve.set_logger(server_logger)
-            serve.logger = server_logger
-        except Exception:
-            # This happens if logging has already been set up for wptserve
-            pass
-
-    def setup_routes(self):
-        for url, paths in self.test_paths.iteritems():
-            if url == "/":
-                continue
-
-            path = paths["tests_path"]
-            url = "/%s/" % url.strip("/")
-
-            for (method,
-                 suffix,
-                 handler_cls) in [(serve.any_method,
-                                   b"*.py",
-                                   serve.handlers.PythonScriptHandler),
-                                  (b"GET",
-                                   "*.asis",
-                                   serve.handlers.AsIsHandler),
-                                  (b"GET",
-                                   "*",
-                                   serve.handlers.FileHandler)]:
-                route = (method, b"%s%s" % (str(url), str(suffix)), handler_cls(path, url_base=url))
-                serve.routes.insert(-3, route)
-
-        if "/" not in self.test_paths:
-            serve.routes = serve.routes[:-3]
-
-    def copy_required_files(self):
-        logger.info("Placing required files in server environment.")
-        for source, destination, copy_if_exists in self.required_files:
-            source_path = os.path.join(here, source)
-            dest_path = os.path.join(serve_path(self.test_paths), destination, os.path.split(source)[1])
-            dest_exists = os.path.exists(dest_path)
-            if not dest_exists or copy_if_exists:
-                if dest_exists:
-                    backup_path = dest_path + ".orig"
-                    logger.info("Backing up %s to %s" % (dest_path, backup_path))
-                    self.files_to_restore.append(dest_path)
-                    shutil.copy2(dest_path, backup_path)
-                logger.info("Copying %s to %s" % (source_path, dest_path))
-                shutil.copy2(source_path, dest_path)
-
-    def ensure_started(self):
-        # Pause for a while to ensure that the server has a chance to start
-        time.sleep(2)
-        for scheme, servers in self.servers.iteritems():
-            for port, server in servers:
-                if self.test_server_port:
-                    s = socket.socket()
-                    try:
-                        s.connect((self.config["host"], port))
-                    except socket.error:
-                        raise EnvironmentError(
-                            "%s server on port %d failed to start" % (scheme, port))
-                    finally:
-                        s.close()
-
-                if not server.is_alive():
-                    raise EnvironmentError("%s server on port %d failed to start" % (scheme, port))
-
-    def restore_files(self):
-        for path in self.files_to_restore:
-            os.unlink(path)
-            if os.path.exists(path + ".orig"):
-                os.rename(path + ".orig", path)
-
-
-class LogThread(threading.Thread):
-    def __init__(self, queue, logger, level):
-        self.queue = queue
-        self.log_func = getattr(logger, level)
-        threading.Thread.__init__(self, name="Thread-Log")
-        self.daemon = True
-
-    def run(self):
-        while True:
-            try:
-                msg = self.queue.get()
-            except (EOFError, IOError):
-                break
-            if msg is None:
-                break
-            else:
-                self.log_func(msg)
-
-
-class LoggingWrapper(StringIO):
-    """Wrapper for file like objects to redirect output to logger
-    instead"""
-
-    def __init__(self, queue, prefix=None):
-        self.queue = queue
-        self.prefix = prefix
-
-    def write(self, data):
-        if isinstance(data, str):
-            data = data.decode("utf8")
-
-        if data.endswith("\n"):
-            data = data[:-1]
-        if data.endswith("\r"):
-            data = data[:-1]
-        if not data:
-            return
-        if self.prefix is not None:
-            data = "%s: %s" % (self.prefix, data)
-        self.queue.put(data)
-
-    def flush(self):
-        pass
+    logger = wptlogging.setup(*args, **kwargs)
 
 def get_loader(test_paths, product, debug=False, **kwargs):
     run_info = wpttest.get_run_info(kwargs["run_info"], product, debug=debug)
@@ -327,8 +60,7 @@ def get_loader(test_paths, product, debug=False, **kwargs):
     return run_info, test_loader
 
 def list_test_groups(test_paths, product, **kwargs):
-
-    do_delayed_imports(serve_path(test_paths))
+    env.do_delayed_imports(logger, test_paths)
 
     run_info, test_loader = get_loader(test_paths, product,
                                        **kwargs)
@@ -338,7 +70,8 @@ def list_test_groups(test_paths, product, **kwargs):
 
 
 def list_disabled(test_paths, product, **kwargs):
-    do_delayed_imports(serve_path(test_paths))
+    env.do_delayed_imports(logger, test_paths)
+
     rv = []
 
     run_info, test_loader = get_loader(test_paths, product,
@@ -350,47 +83,27 @@ def list_disabled(test_paths, product, **kwargs):
     print json.dumps(rv, indent=2)
 
 
-def get_ssl_kwargs(**kwargs):
-    if kwargs["ssl_type"] == "openssl":
-        args = {"openssl_binary": kwargs["openssl_binary"]}
-    elif kwargs["ssl_type"] == "pregenerated":
-        args = {"host_key_path": kwargs["host_key_path"],
-                "host_cert_path": kwargs["host_cert_path"],
-                 "ca_cert_path": kwargs["ca_cert_path"]}
-    else:
-        args = {}
-    return args
+def get_pause_after_test(test_loader, **kwargs):
+    total_tests = sum(len(item) for item in test_loader.tests.itervalues())
+    if kwargs["pause_after_test"] is None:
+        if kwargs["repeat"] == 1 and total_tests == 1:
+            return True
+        return False
+    return kwargs["pause_after_test"]
 
-def serve_path(test_paths):
-    return test_paths["/"]["tests_path"]
 
 def run_tests(config, test_paths, product, **kwargs):
-    logging_queue = None
-    logging_thread = None
-    original_stdio = (sys.stdout, sys.stderr)
-    test_queues = None
-
-    try:
-        if not kwargs["no_capture_stdio"]:
-            logging_queue = Queue()
-            logging_thread = LogThread(logging_queue, logger, "info")
-            sys.stdout = LoggingWrapper(logging_queue, prefix="STDOUT")
-            sys.stderr = LoggingWrapper(logging_queue, prefix="STDERR")
-            logging_thread.start()
-
-        do_delayed_imports(serve_path(test_paths))
+    with wptlogging.CaptureIO(logger, not kwargs["no_capture_stdio"]):
+        env.do_delayed_imports(logger, test_paths)
 
         (check_args,
          browser_cls, get_browser_kwargs,
          executor_classes, get_executor_kwargs,
          env_options) = products.load_product(config, product)
 
+        ssl_env = env.ssl_env(logger, **kwargs)
+
         check_args(**kwargs)
-
-        ssl_env_cls = sslutils.environments[kwargs["ssl_type"]]
-        ssl_env = ssl_env_cls(logger, **get_ssl_kwargs(**kwargs))
-
-        unexpected_total = 0
 
         if "test_loader" in kwargs:
             run_info = wpttest.get_run_info(kwargs["run_info"], product, debug=False)
@@ -409,12 +122,14 @@ def run_tests(config, test_paths, product, **kwargs):
 
         logger.info("Using %i client processes" % kwargs["processes"])
 
-        with TestEnvironment(test_paths,
-                             ssl_env,
-                             env_options) as test_environment:
+        unexpected_total = 0
+
+        with env.TestEnvironment(test_paths,
+                                 ssl_env,
+                                 env_options) as test_environment:
             try:
                 test_environment.ensure_started()
-            except TestEnvironmentError as e:
+            except env.TestEnvironmentError as e:
                 logger.critical("Error starting test environment: %s" % e.message)
                 raise
 
@@ -469,22 +184,6 @@ def run_tests(config, test_paths, product, **kwargs):
                 unexpected_total += unexpected_count
                 logger.info("Got %i unexpected results" % unexpected_count)
                 logger.suite_end()
-    except KeyboardInterrupt:
-        if test_queues is not None:
-            for queue in test_queues.itervalues():
-                queue.cancel_join_thread()
-    finally:
-        if test_queues is not None:
-            for queue in test_queues.itervalues():
-                queue.close()
-        sys.stdout, sys.stderr = original_stdio
-        if not kwargs["no_capture_stdio"] and logging_queue is not None:
-            logger.info("Closing logging queue")
-            logging_queue.put(None)
-            if logging_thread is not None:
-                logging_thread.join(10)
-            logging_queue.close()
-            logger.info("queue closed")
 
     return unexpected_total == 0
